backend/requirements.txt
fastapi==0.115.8
uvicorn==0.41.0
groq==0.16.0
langgraph==0.2.73
pydantic==2.10.6
python-dotenv==1.0.1
